now take into account the fill-in in R, and updated merge strategy
New merge strategy:
in the main loop, we now only increase cwmax (k=merges are considered
only for k <= cwmax) when the number of merges is below a given ratio
of the initial number of merges for this value of cwmax. This ensures
we don't increase cwmax too early.

In particular, if this ratio is < 1, we perform at least two passes
for each value of cwmax.