disable the b256 build (it's unused anyway)
The immediate goal is to get some exposure for the variable-width code.

In truth, we should probably opt to compile everything, and allow
arith_generic::instance to select an instance not only based on the
width, but also possible external bits of information such as a command
line parameter (a bit like what we have with mm_impl).