expand polynomial<> to cover other types.
It's not complete at this point, but at least the tests pass.

There are quite a few subtle points that should be tested better than
what is currently done. In particular, some care is taken so that (for
example) a polynomial<cxx_mpfr> evaluated at an int and a
polynomial<int> evaluated at a cxx_mpfr both return a cxx_mpfr,
hopefully with a correctly defined precision. This works with some
template trickery.

There are also a few constructs that should be refactored (templating on
template args is probably the least flexible SFINAE strategy, we should
do away with it).