Fix polynomial<cxx_mpfr>
When T is a contextual type such as cxx_mpfr, make polynomial<T> inherit
from a traits object that has the context (in that case, precision)
information.

Note that the test_resultant function is not extraordinarily relevant.

We killed the number_traits::from_number_literal interface, because it
must have access to the context if we happen to have one.